rf/changes proto from repeated uint32 to bytes

+++ src/pb.rs
@@ -21,8 +21,8 @@ pub struct GetFileRequest {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct GetFileResponse {
-    #[prost(uint32, repeated, tag = "1")]
-    pub chunk: ::prost::alloc::vec::Vec<u32>,
+    #[prost(bytes = "vec", tag = "1")]
+    pub chunk: ::prost::alloc::vec::Vec<u8>,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
